requirements.txt
netmiko
pyyaml
networkx
matplotlib